requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
filelock==3.4.2
Flask==2.0.2
Flask-Compress==1.10.1
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.5
packaging==21.3
plotly==5.5.0
praw==7.5.0
prawcore==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scipy==1.7.3
six==1.16.0
tenacity==8.0.1
tokenizers==0.10.3
torch==1.10.1
tqdm==4.62.3
transformers==4.15.0
typing-extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.2
zipp==3.7.0
